requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
gunicorn==21.2.0
python-multipart==0.0.6
python-dotenv==1.0.0
openai==1.51.0
httpx==0.27.0